fix: fixed the slider and made it render
