+++ src/pages/main-search/mainFilter.ts
@@ -1,5 +1,6 @@
 import products from '../../products.json';
 import MainPage, { Product } from '../main-page/main';
+import * as noUiSlider from '../../../node_modules/nouislider';
 
 class FilterProducts {
   private id: string;
@@ -114,7 +115,7 @@ class FilterProducts {
       if (this.filterCheckCategory.length === 0) {
         arr = products.filter(item => arrCheckedRus.includes(item.brand));
       }
-      console.log(arr.filter(item => arrCheckedRus.includes(item.brand)))
+      // console.log(arr.filter(item => arrCheckedRus.includes(item.brand)))
       this.mainPage.render(arr.filter(item => arrCheckedRus.includes(item.brand)), 'search');
       this.arrSort = arr;
       return arr.filter(item => arrCheckedRus.includes(item.brand));
@@ -155,6 +156,95 @@ class FilterProducts {
       case '10kingdom': return 'Десятое королевство';
       case 'zvezda': return 'Zvezda';
       default: return 'Нет такого бренда'
+    }
+  }
+// g
+
+  noSlider (): void {
+    const sliderPrice = document.getElementById('slider-price') as HTMLDivElement;
+
+    const sliderPr = noUiSlider.create(sliderPrice, {
+      start: [0, 263],
+      connect: true,
+      range: {
+        min: 0,
+        max: 263
+      },
+      tooltips: false,
+      format: {
+        from: function (value) {
+          return parseInt(value);
+        },
+        to: function (value) {
+          return Math.floor(value);
+        }
+      }
+    });
+
+    const rangePriceFrom = document.getElementById('input-price-from') as HTMLInputElement;
+    const rangePriceTo = document.getElementById('input-price-to') as HTMLInputElement;
+
+    if (sliderPrice instanceof HTMLDivElement) {
+      sliderPr.on('update', (values, handler) => {
+        const allValues: number[] | string[] = [];
+        allValues[handler] = values[handler].toString();
+        rangePriceFrom.value = values[0].toString()
+        rangePriceTo.value = values[1].toString()
+
+        const arrPrice: Product[] = [];
+        this.arrSort.forEach(element => {
+          if (element.price >= Number(rangePriceFrom.value) && element.price <= Number(rangePriceTo.value)) {
+            arrPrice.push(element);
+          }
+        })
+        // this.arrSort = arrPrice;
+        this.mainPage.render(arrPrice, 'search')
+        // console.log('price', arrPrice);
+      })
+    }
+
+    /* noUiSlider на количесвто товара */
+
+    const sliderAmount = document.getElementById('slider-amount') as HTMLDivElement;
+
+    const sliderAm = noUiSlider.create(sliderAmount, {
+      start: [0, 16],
+      connect: true,
+      range: {
+        min: 0,
+        max: 16
+      },
+      tooltips: false,
+      format: {
+        from: function (value) {
+          return parseInt(value);
+        },
+        to: function (value) {
+          return Math.floor(value);
+        }
+      }
+    });
+    const rangeAmountFrom = document.getElementById('input-amount-from') as HTMLInputElement;
+    const rangeAmountTo = document.getElementById('input-amount-to') as HTMLInputElement;
+
+    if (sliderAmount instanceof HTMLDivElement) {
+      sliderAm.on('update', (values, handler) => {
+        const allValues: number[] | string[] = [];
+        // console.log(values[handler].toString());
+        allValues[handler] = values[handler].toString();
+        rangeAmountFrom.value = values[0].toString();
+        rangeAmountTo.value = values[1].toString();
+
+        const arrAmount: Product[] = [];
+        this.arrSort.forEach(element => {
+          if (element.amount >= Number(rangeAmountFrom.value) && element.amount <= Number(rangeAmountTo.value)) {
+            arrAmount.push(element);
+          }
+        })
+        // this.arrSort = arrAmount;
+        this.mainPage.render(arrAmount, 'search')
+        // console.log('amount', arrAmount);
+      })
     }
   }
 }
